requirements.txt
argcomplete==1.9.2
base58==0.2.4
boto3==1.4.8
botocore==1.8.11
certifi==2017.11.5
cfn-flip==0.3.0
chardet==3.0.4
click==6.7
Django==1.11.7
djangorestframework==3.7.3
docutils==0.14
durationpy==0.5
future==0.16.0
futures==3.1.1
hjson==3.0.1
idna==2.6
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
placebo==0.8.1
psycopg2==2.7.3.2
python-dateutil==2.6.1
python-decouple==3.1
python-slugify==1.2.4
pytz==2017.3
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.12
six==1.11.0
toml==0.9.3
tqdm==4.19.1
troposphere==2.1.2
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12
wsgi-request-logger==0.4.6
zappa==0.45.1
